Cleaned up and restructured SqlFilter trait and its base impls
