+++ src/models/user.rs
@@ -2,6 +2,7 @@ use crate::traits::model::Model;
 use chrono::NaiveDateTime;
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
+use sqlx_filter_macro::sql_filter;
 use uuid::Uuid;
 
 #[derive(
@@ -22,5 +23,21 @@ impl Model for User {
 
     fn get_id(&self) -> Option<Self::Id> {
         self.ext_id
+    }
+}
+
+sql_filter! {
+    #[derive(Default, Debug, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
+    pub struct UserFilter {
+        SELECT * FROM users
+        WHERE
+            ?id = i64
+            AND ext_id = Uuid
+            AND ?username = String
+            OR ?email = String
+            AND ?created_at as created_before < NaiveDateTime
+            AND ?created_at as created_after > NaiveDateTime
+            AND ?updated_at as updated_before < NaiveDateTime
+            AND ?updated_at as updated_after > NaiveDateTime
     }
 }
