--- src/traits/sql_filter/equals.rs
@@ -1,3 +1,0 @@
-use crate::traits::sql_operator;
-
-sql_operator!(pub Equals, "=");
