--- src/traits/sql_filter/i_like.rs
@@ -1,3 +1,0 @@
-use crate::traits::sql_operator;
-
-sql_operator!(pub ILike<String>, "ILIKE");
