--- src/traits/sql_filter/less_than_or_equal.rs
@@ -1,3 +1,0 @@
-use crate::traits::sql_operator;
-
-sql_operator!(pub LessThanOrEqual, "<=");
