--- src/traits/sql_filter/not_in.rs
@@ -1,3 +1,0 @@
-use crate::traits::sql_operator;
-
-sql_operator!(pub NotInValues[], "NOT IN");
