+++ src/utils/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod api_scope;
 pub(crate) mod batch;
+mod db;
 pub(crate) mod middleware_macros;
 pub mod mod_def;
-pub mod sql_filter;
 
 pub(crate) use api_scope::api_scope;
