+++ src/utils/mod_def.rs
@@ -1,8 +1,8 @@
 #[macro_export]
 macro_rules! mod_def {
-    {$vis:vis mod $ident:ident $(;)?} => {
-        $vis mod $ident;
+    {$($vis:vis mod $ident:ident $(;)?)+} => {
+        $($vis mod $ident;
         #[allow(unused_imports)]
-        $vis use $ident::*;
+        $vis use $ident::*;)+
     };
 }
